CSA Plan
This program in Python was design to model the layout for a Community Supported Agricultural (CSA) garden. The purpose of this program was to organize the plants grown and the layout of the garden to optimize profits, but due to time constraints only a basic feasible solution could be determined. I worked alone on this project.